3-DivingDeepIntoTwitterAPI/pyproject.toml
[tool.poetry]
name = "3-divingdeepintotwitterapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
tweepy = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
